+++ src/main/kotlin/internals/instructions/controlFlow/jz.kt
@@ -15,7 +15,7 @@ import vm
  */
 fun ControlFlow.jz(targetAddress: Long): Any = try {
 	if (registers.read(intelNames[IntelRegisters.ZF]) == 0L) {
-		vm.pc = targetAddress.toInt()
+		vm.pc = targetAddress
 	} else {
 		// I'm not sure why I need this else block
 	}
